Flytter regler for beregning av barnepensjon til beregnings-appen.
EY-1543
