+++ apps/etterlatte-beregning/src/main/kotlin/beregning/regler/trygdetidsfaktor/Trygdetidsfaktor.kt
@@ -4,9 +4,9 @@ import no.nav.etterlatte.libs.regler.Regel
 import no.nav.etterlatte.libs.regler.RegelMeta
 import no.nav.etterlatte.libs.regler.RegelReferanse
 import no.nav.etterlatte.libs.regler.benytter
-import no.nav.etterlatte.libs.regler.beregning.AvdoedForelder
-import no.nav.etterlatte.libs.regler.beregning.BarnepensjonGrunnlag
+import no.nav.etterlatte.libs.regler.beregning.barnepensjon1967.AvdoedForelder
 import no.nav.etterlatte.libs.regler.beregning.barnepensjon1967.BP_1967_DATO
+import no.nav.etterlatte.libs.regler.beregning.barnepensjon1967.BarnepensjonGrunnlag
 import no.nav.etterlatte.libs.regler.definerKonstant
 import no.nav.etterlatte.libs.regler.finnFaktumIGrunnlag
 import no.nav.etterlatte.libs.regler.med
