+++ libs/etterlatte-regler/src/test/kotlin/regler/VelgNyesteGyldigeRegelTest.kt
@@ -2,7 +2,6 @@ package no.nav.etterlatte.libs.regler
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
-import no.nav.etterlatte.libs.regler.beregning.toDoRegelReferanse
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
