+++ libs/etterlatte-regler/src/test/kotlin/regler/VisitorTest.kt
@@ -3,7 +3,6 @@ package no.nav.etterlatte.libs.regler
 import io.kotest.matchers.collections.shouldContainExactlyInAnyOrder
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
-import no.nav.etterlatte.libs.regler.beregning.toDoRegelReferanse
 import org.junit.jupiter.api.Test
 import java.time.LocalDate
